Revert "Fixed Crash due to null Activity in Life Cycle Event Listener in Android Module"
