+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@rudderstack/rudder-sdk-react-native",
-  "version": "1.5.1",
+  "version": "1.5.0",
   "description": "Rudder React Native SDK",
   "main": "index.ts",
   "keywords": [
